\name{mypkg02-package}
\alias{mypkg02-package}
\alias{mypkg02}
\docType{package}
\title{
Calculate Sum using R and C
}
\description{
Calculate summation of  numeric vector by R and C
}
\details{
\tabular{ll}{
Package: \tab mypkg02\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-06\cr
License: \tab GPL-2 \cr
}
An overview of how to use the package, including the most important fn
}
\author{
M. R. Karim <mrezakarim@yahoo.com>

Maintainer: M. R. Karim <mrezakarim@yahoo.com>
}
